Moves schema-related code to `schema.py`
