+++ cerberus/tests/__init__.py
@@ -1,4 +1,4 @@
-from ..cerberus import Validator, SchemaError, DocumentError, errors
+from .. import Validator, SchemaError, DocumentError, errors
 
 import sys
 if sys.version_info >= (2, 7):
